+++ src/map.h
@@ -10,7 +10,7 @@ struct MapCoords { int x, z; };
 struct Map {
 	char path[1024];
 	char name[30];   // Must fit in map chooser and map delete dialog
-	float sortkey;
+	double sortkey;
 	bool custom;  // whether path starts with "custom_maps"
 	struct Wall walls[MAX_WALLS];
 	int nwalls;
